requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2
django-auditlog==3.2.1
django-autocomplete-light==3.12.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
